+++ app/src/main/java/com/art241111/kprizes/ui/tintGame/tintScreen/CatTextTintScreenVM.kt
@@ -1,8 +1,6 @@
 package com.art241111.kprizes.ui.tintGame.tintScreen
 
-import androidx.compose.runtime.State
-import androidx.lifecycle.ViewModel
-import com.art241111.kprizes.ui.catHelper.CatTextVM
+import com.art241111.kprizes.ui.catHelper.OpenCatTextVM
 
 /**
  * View model for updating phrases on the tint screen
@@ -10,43 +8,9 @@ import com.art241111.kprizes.ui.catHelper.CatTextVM
  * @author Created by Artem Gerasimov (gerasimov.av.dev@gmail.com).
  */
 
-class CatTextTintScreenVM : ViewModel() {
-    private val catTextVM = CatTextVM()
-
-    /**
-     * The state that changes when the text changes.
-     * Required for animating text changes in the assistant cat.
-     */
-    val isUpdate: State<Boolean> = catTextVM.isUpdate
-
-    /**
-     * The text that the assistant cat is currently saying.
-     */
-    val text: State<String> = catTextVM.text
-
-    /**
-     * Starting the text update for the assistant cat.
-     */
-    private var isRun = false
-    fun startUpdate() {
-        if (!isRun) {
-            isRun = true
-
-            catTextVM.startUpdate(
-                listOf(
-                    "Наклоняйте планшет, чтобы управлять роботом",
-                    // "Давай сыграем в игру!",
-                    // "Вытяни игрушку с помощью промышленного робота "
-                )
-            )
-        }
-    }
-
-    /**
-     * Stopping text updates.
-     */
-    fun stopUpdate() {
-        isRun = false
-        catTextVM.stopUpdate()
-    }
+class CatTextTintScreenVM : OpenCatTextVM() {
+    override val possibleText: List<String>
+        get() = listOf(
+            "Наклоняйте планшет, чтобы управлять роботом",
+        )
 }
